fix(compare): move comparison logic to a new file and simplify equality check
